refactor: 예외 처리 코드 number, numbers로 파일 분리
